+++ CHANGELOG.md
@@ -1,7 +1,15 @@
 # Change Log for SD.Next
 
-## Update for 2025-02-07  
+## Update for 2025-02-08  
 
+- **Hints**  
+  - added/updated 100+ ui hints!  
+  - [hints](https://github.com/vladmandic/wiki/Hints) documentation and contribution guide  
+- **Localization**  
+  - full ui localization!  
+    english, croatian, spanish, french, italian, portuguese, chinese, japanese, korean, russian  
+  - set in *settings -> user interface -> language*
+  - [localization](https://github.com/vladmandic/wiki/Locale) documentation  
 - **Torch**:  
   - for **zluda** set default to `torch==2.6.0+cu126`  
 - **Other**: